change type int for idx

+++ 0x1E-search_algorithms/0-linear.c
@@ -11,13 +11,13 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-    size_t idx;
+    unsigned int idx;
 
     if (array != NULL)
     {
         for(idx = 0; idx < size; idx++)
         {
-            printf("Value checked array[%lu] = [%d]\n", idx, array[idx]);
+            printf("Value checked array[%d] = [%d]\n", idx, array[idx]);
             if (array[idx] == value)
                 return (idx);
         }
